+++ CHANGELOG.md
@@ -1,2 +1,6 @@
-* Added config option:
-  * Breadcrumbs On World Map
+* Add layer info display
+* Added config options:
+  * Ghost Seek Breadcrumb Colors
+    * Match With Action Bar Pings
+    * ... and 5 options to customize each ping color
+* Fixed `Clear Breadcrumbs On Find` sometimes not working